+++ WebApi/Controllers/Api/Test/TestController.cs
@@ -4,7 +4,7 @@ using System.Web.Http;
 namespace forCrowd.WealthEconomy.WebApi.Controllers.Api.Test
 {
     /// <summary>
-    /// A dummy controller to test IHttpAction results 
+    /// A dummy controller to test some functions and results
     /// </summary>
     [AllowAnonymous]
     [RoutePrefix("api/Test")]
@@ -17,18 +17,33 @@ namespace forCrowd.WealthEconomy.WebApi.Controllers.Api.Test
             return Ok(string.Empty);
         }
 
+        //[Route("SendTestEmail")]
+        //[HttpGet]
+        //public async System.Threading.Tasks.Task<IHttpActionResult> SendTestEmail()
+        //{
+        //    var message = new Microsoft.AspNet.Identity.IdentityMessage();
+        //    message.Subject = "email test";
+        //    message.Body = "email test";
+        //    message.Destination = "local@forcrowd.org";
+
+        //    var emailService = new Facade.EmailService();
+        //    await emailService.SendAsync(message);
+
+        //    return Ok(message);
+        //}
+
+        [Route("ThrowException")]
+        [HttpGet]
+        public void ThrowException()
+        {
+            throw new InvalidOperationException("Test exception");
+        }
+
         [Route("UnauthorizedResult")]
         [HttpGet]
         public IHttpActionResult UnauthorizedResult()
         {
             return Unauthorized();
-        }
-
-        [Route("SomeException")]
-        [HttpGet]
-        public void SomeException()
-        {
-            throw new InvalidOperationException("Result test");
         }
     }
 }