+++ ng2Client/app/components/account/register.component.ts
@@ -3,12 +3,12 @@ import { Router } from "@angular/router";
 
 import { DataService } from "../../services/data.service";
 import { Logger } from "../../services/logger.service";
-import { AppSettings } from "../../settings/app-settings";
+import { Settings } from "settings";
 
 @Component({
     moduleId: module.id,
     selector: "register",
-    templateUrl: "register.component.html?v=" + AppSettings.version
+    templateUrl: "register.component.html?v=" + Settings.version
 })
 export class RegisterComponent implements OnInit {