+++ ng2Client/app/rxjs-extensions.ts
@@ -1,5 +1,6 @@
 // Observable class extensions
 import "rxjs/add/observable/empty";
+import "rxjs/add/observable/forkJoin";
 import "rxjs/add/observable/fromPromise";
 import "rxjs/add/observable/of";
 import "rxjs/add/observable/timer";
@@ -13,4 +14,5 @@ import "rxjs/add/operator/filter";
 import "rxjs/add/operator/finally";
 import "rxjs/add/operator/map";
 import "rxjs/add/operator/mergeMap";
+import "rxjs/add/operator/share";
 import "rxjs/add/operator/switchMap";